+++ lib/src/subtypes/guilds/thieves_guild/thieves_guild.dart
@@ -1,3 +1,5 @@
+import '../../emblems/default/default_emblem_type.dart';
+import '../../emblems/emblem_type.dart';
 import '../../../generators/base/generator.dart';
 import '../../../generators/base/list_item_generator.dart';
 import '../../../generators/base/multiple_generator.dart';
@@ -66,4 +68,7 @@ class ThievesGuild implements GuildType {
 
   @override
   IGenerator<String> getSpecialtyGenerator() => ThievesSpecialtyGenerator();
+
+  @override
+  EmblemType getEmblemType() => DefaultEmblemType();
 }